requirements.txt
flask
line-bot-sdk
notebooklm-py
python-dotenv
asyncio
